fix(T2): Код разделен по единицам трансляции
