+++ miheev.egor/T2/scopeguard.cpp
@@ -1,0 +1,15 @@
+#include "scopeguard.hpp"
+
+miheev::iofmtguard::iofmtguard(std::basic_ios< char >& s):
+  s_(s),
+  fill_(s.fill()),
+  precision_(s.precision()),
+  fmt_(s.flags())
+{}
+
+miheev::iofmtguard::~iofmtguard()
+{
+  s_.fill(fill_);
+  s_.precision(precision_);
+  s_.flags(fmt_);
+}
