+++ miheev.egor/T2/scopeguard.hpp
@@ -1,0 +1,21 @@
+#ifndef SCOPEGUARD_HPP
+#define SCOPEGUARD_HPP
+
+#include <iostream>
+
+namespace miheev
+{
+  class iofmtguard
+  {
+  public:
+    iofmtguard(std::basic_ios< char >& s);
+    ~iofmtguard();
+  private:
+    std::basic_ios< char >& s_;
+    char fill_;
+    std::streamsize precision_;
+    std::basic_ios< char >::fmtflags fmt_;
+  };
+}
+
+#endif
